+++ backend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.14.0](https://github.com/skang10/signalyst/compare/signalyst-backend-v0.13.0...signalyst-backend-v0.14.0) (2026-06-09)
+
+
+### Features
+
+* add follow-up analysis and refresh session UI ([#49](https://github.com/skang10/signalyst/issues/49)) ([a34eb66](https://github.com/skang10/signalyst/commit/a34eb664fbf3300e964998f375bf17cea4a6cc36))
+* add FollowUpAgent and route FOLLOW_UP-stage chat to it (PR 4b) ([#47](https://github.com/skang10/signalyst/issues/47)) ([e0a4c77](https://github.com/skang10/signalyst/commit/e0a4c77a516f48df9ec8991ff27ffdcde99beb40))
+
+## [0.13.0](https://github.com/skang10/signalyst/compare/signalyst-backend-v0.12.0...signalyst-backend-v0.13.0) (2026-06-07)
+
+
+### Features
+
+* add ExplanationAgent and wire EXPLAINING into the pipeline ([#41](https://github.com/skang10/signalyst/issues/41)) ([542435b](https://github.com/skang10/signalyst/commit/542435b907705413cd98fb79e470bb34064f8a9e))
+* move session tabs to a vertical sidebar and add an editable Config tab ([#39](https://github.com/skang10/signalyst/issues/39)) ([4e1bf7b](https://github.com/skang10/signalyst/commit/4e1bf7bce1a69904f0c3d547576e6383267fa1d9))
+
 ## [0.12.0](https://github.com/skang10/signalyst/compare/signalyst-backend-v0.11.0...signalyst-backend-v0.12.0) (2026-06-07)